Struct cardano_legacy_address::crc32::Crc32

pub struct Crc32(u32);

structure to compute the CRC32 of chunks of bytes.

This structure allows implements the Write trait making it easier to compute the crc32 of a stream.

Tuple Fields
0: u32
Implementations

impl Crc32

pub fn new() -> Self

initialise a new CRC32 state

pub fn update<'a, I>(&'a mut self, bytes: I) -> &mut Selfwhere I: IntoIterator<Item = &'a u8>,

update the CRC32 with the given bytes.

beware that the order in which you update the Crc32 matter.

pub fn finalize(self) -> u32

finalize the CRC32, recovering the computed value